researchflow-production-main/ci-debugging-agent/requirements.txt
# CI Debugging Agent Requirements
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0

# Composio Integration
composio>=0.5.0
composio_langchain>=0.8.0

# GitHub Integration
PyGithub>=2.1.0
gitpython>=3.1.40

# Code Analysis
tree-sitter>=0.22.0
tree-sitter-typescript>=0.20.0
tree-sitter-python>=0.21.0
ast-grep>=0.12.0
pylint>=3.0.0
mypy>=1.8.0

# Docker Analysis
docker>=7.0.0
dockerfile-parse>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
rich>=13.7.0
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
